feat(apis_metainfo): move signal receiver to correct app
Closes: #2006
